Move library to own package and install that
