+++ workflow/scripts/build_raw_blind.py
@@ -16,9 +16,9 @@ from pathlib import Path
 import numexpr as ne
 import numpy as np
 from dbetto.catalog import Props
+from legenddataflow.log import build_log
 from legendmeta import LegendMetadata, TextDB
 from lgdo import lh5
-from library.log import build_log
 
 argparser = argparse.ArgumentParser()
 argparser.add_argument("--input", help="input file", type=str)
