+++ workflow/scripts/build_tcm.py
@@ -6,7 +6,7 @@ import numpy as np
 from daq2lh5.orca import orca_flashcam
 from dbetto import TextDB
 from dbetto.catalog import Props
-from library.log import build_log
+from legenddataflow.log import build_log
 from pygama.evt.build_tcm import build_tcm
 
 argparser = argparse.ArgumentParser()
