+++ workflow/scripts/pars_dsp_dplms_geds.py
@@ -8,9 +8,9 @@ import lgdo.lh5 as lh5
 import numpy as np
 from dbetto import TextDB
 from dbetto.catalog import Props
+from legenddataflow.log import build_log
 from legendmeta import LegendMetadata
 from lgdo import Array, Table
-from library.log import build_log
 from pygama.pargen.dplms_ge_dict import dplms_ge_dict
 
 argparser = argparse.ArgumentParser()
