+++ workflow/scripts/pars_dsp_eopt_geds.py
@@ -11,8 +11,8 @@ import sklearn.gaussian_process.kernels as ker
 from dbetto import TextDB
 from dbetto.catalog import Props
 from dspeed.units import unit_registry as ureg
+from legenddataflow.log import build_log
 from legendmeta import LegendMetadata
-from library.log import build_log
 from pygama.math.distributions import hpge_peak
 from pygama.pargen.dsp_optimize import (
     BayesianOptimizer,
