+++ workflow/scripts/pars_dsp_evtsel_geds.py
@@ -12,8 +12,8 @@ import pygama.math.histogram as pgh
 import pygama.pargen.energy_cal as pgc
 from dbetto import TextDB
 from dbetto.catalog import Props
+from legenddataflow.log import build_log
 from legendmeta import LegendMetadata
-from library.log import build_log
 from pygama.pargen.data_cleaning import generate_cuts, get_keys, get_tcm_pulser_ids
 from pygama.pargen.dsp_optimize import run_one_dsp
