+++ workflow/scripts/pars_hit_qc.py
@@ -11,10 +11,10 @@ from pathlib import Path
 import numpy as np
 from dbetto import TextDB
 from dbetto.catalog import Props
+from legenddataflow.convert_np import convert_dict_np_to_float
+from legenddataflow.log import build_log
 from legendmeta import LegendMetadata
 from lgdo.lh5 import ls
-from library.convert_np import convert_dict_np_to_float
-from library.log import build_log
 from pygama.pargen.data_cleaning import (
     generate_cut_classifiers,
     get_keys,
