+++ workflow/scripts/pars_pht_aoecal.py
@@ -13,9 +13,9 @@ import numpy as np
 import pandas as pd
 from dbetto import TextDB
 from dbetto.catalog import Props
+from legenddataflow.FileKey import ChannelProcKey, ProcessingFileKey
+from legenddataflow.log import build_log
 from legendmeta import LegendMetadata
-from library.FileKey import ChannelProcKey, ProcessingFileKey
-from library.log import build_log
 from pygama.pargen.AoE_cal import *  # noqa: F403
 from pygama.pargen.AoE_cal import CalAoE, Pol1, SigmaFit, aoe_peak
 from pygama.pargen.data_cleaning import get_tcm_pulser_ids
