+++ workflow/scripts/pars_pht_fast.py
@@ -10,9 +10,9 @@ import numpy as np
 import pandas as pd
 from dbetto import TextDB
 from dbetto.catalog import Props
+from legenddataflow.FileKey import ChannelProcKey, ProcessingFileKey
+from legenddataflow.log import build_log
 from legendmeta import LegendMetadata
-from library.FileKey import ChannelProcKey, ProcessingFileKey
-from library.log import build_log
 from pars_pht_aoecal import run_aoe_calibration
 from pars_pht_lqcal import run_lq_calibration
 from pars_pht_partcal import calibrate_partition
