+++ workflow/scripts/pars_tcm_pulser.py
@@ -6,8 +6,8 @@ import lgdo.lh5 as lh5
 import numpy as np
 from dbetto import TextDB
 from dbetto.catalog import Props
+from legenddataflow.log import build_log
 from legendmeta import LegendMetadata
-from library.log import build_log
 from pygama.pargen.data_cleaning import get_tcm_pulser_ids
 
 argparser = argparse.ArgumentParser()
